car_diagnostic_agent/pyproject.toml

[project]
name = "car-diagnostic-agent"
version = "0.1.0"
description = "A2A Car Diagnostic Agent"
dependencies = [
    "a2a-sdk[http-server]",
    "langchain-google-genai",
    "python-dotenv",
    "uvicorn",
    "pyserial>=3.5",
    "obd>=0.7.1",
    "asyncio-mqtt>=0.13.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "azure-cognitiveservices-speech>=1.35.0",
]


